requirements.txt
aspose-slides
matplotlib
plotly
pandas
kaleido
